+++ tests/test_predict_on_batch.py
@@ -60,7 +60,7 @@ class TestKernelParameterValidation:
         """An invalid parameter raise an error."""
         X, y = synthetic_data
         im = ImpactModel(lm, rng_key=random.key(42), vi=vi)
-        im.fit(X=X, y=y, progress=False)
+        im.fit(X=X, y=y, batch_size=3)
         with pytest.raises(TypeError):
             im.predict_on_batch(X=X, y=y)
 
@@ -73,7 +73,7 @@ class TestKernelParameterValidation:
         """Extra parameters not present in the kernel raise an error."""
         X, y = synthetic_data
         im = ImpactModel(lm, rng_key=random.key(42), vi=vi)
-        im.fit(X=X, y=y, progress=False)
+        im.fit(X=X, y=y, batch_size=3)
         with pytest.raises(TypeError):
             im.predict_on_batch(X=X, y=y, extra=True)
 
@@ -95,7 +95,7 @@ class TestKernelParameterValidation:
             loss=Trace_ELBO(),
         )
         im = ImpactModel(kernel, rng_key=random.key(42), vi=vi)
-        im.fit(X=X, arg=arg, y=y, progress=False)
+        im.fit(X=X, arg=arg, y=y, batch_size=3)
         with pytest.raises(TypeError):
             im.predict_on_batch(X=X)
 
@@ -108,5 +108,5 @@ def test_predict_on_batch_lm_with_kwargs_array(
     """Test the `.predict_on_batch()` method of `ImpactModel`."""
     X, y = synthetic_data
     im = ImpactModel(lm_with_kwargs_array, rng_key=random.key(42), vi=vi)
-    im.fit(X=X, y=y, c=y, progress=False)
+    im.fit(X=X, y=y, c=y, batch_size=3)
     im.predict_on_batch(X=X, c=y)